+++ bsk_rl/src/bsk_rl/act/__init__.py
@@ -1,6 +1,6 @@
 """Actions ``bsk_rl.act`` can be used to add actions to an agent.
 
-To configure the observation, set the ``action_spec`` attribute of a :class:`~bsk_rl.env.scenario.satellites.Satellite`
+To configure the observation, set the ``action_spec`` attribute of a :class:`~bsk_rl.sats.Satellite`
 subclass. For example:
 
 .. code-block:: python
@@ -24,7 +24,7 @@ Use :class:`DiscreteAction` for integer-indexable, discrete actions.
 +----------------------------+---------+-------------------------------------------------------------------------------------------------------+
 | **Action**                 |**Count**| **Description**                                                                                       |
 +----------------------------+---------+-------------------------------------------------------------------------------------------------------+
-| :class:`DiscreteFSWAction` | 1       | Call an arbitrary ``@action`` decorated function in the :class:`~bsk_rl.env.simulation.fsw.FSWModel`. |
+| :class:`DiscreteFSWAction` | 1       | Call an arbitrary ``@action`` decorated function in the :class:`~bsk_rl.sim.fsw.FSWModel`.            |
 +----------------------------+---------+-------------------------------------------------------------------------------------------------------+
 | :class:`Charge`            | 1       | Point the solar panels at the sun.                                                                    |
 +----------------------------+---------+-------------------------------------------------------------------------------------------------------+
@@ -45,18 +45,24 @@ Continuous Actions
 Use :class:`ContinuousAction` for actions with a continuous action space. Currently, satellites
 can only have a single continuous action in their ``action_spec``.
 
-+----------------------------+-------------+-------------------------------------------------------------------------------------------------------+
-| **Action**                 |**Dimension**| **Description**                                                                                       |
-+----------------------------+-------------+-------------------------------------------------------------------------------------------------------+
-| :class:`ImpulsiveThrust`   | 4           | Instantaneously change the satellite's velocity, and drift for some duration.                         |
-+----------------------------+-------------+-------------------------------------------------------------------------------------------------------+
++-----------------------------+-------------+-------------------------------------------------------------------------------------------------------+
+| **Action**                  |**Dimension**| **Description**                                                                                       |
++-----------------------------+-------------+-------------------------------------------------------------------------------------------------------+
+| :class:`ImpulsiveThrust`    | 4           | Instantaneously change the satellite's velocity, and drift for some duration.                         |
++-----------------------------+-------------+-------------------------------------------------------------------------------------------------------+
+| :class:`ImpulsiveThrustHill`| 4           | Like :class:`ImpulsiveThrust`, but specified in the Hill frame of another satellite.                  |
++-----------------------------+-------------+-------------------------------------------------------------------------------------------------------+
 
 
 
 """
 
 from bsk_rl.act.actions import Action
-from bsk_rl.act.continuous_actions import ContinuousAction, ImpulsiveThrust
+from bsk_rl.act.continuous_actions import (
+    ContinuousAction,
+    ImpulsiveThrust,
+    ImpulsiveThrustHill,
+)
 from bsk_rl.act.discrete_actions import (
     Charge,
     Desat,
@@ -81,4 +87,5 @@ __all__ = [
     "Scan",
     "ContinuousAction",
     "ImpulsiveThrust",
+    "ImpulsiveThrustHill",
 ]